\name{abmap}
\alias{abmap}

\title{Example on maps}

\description{
 This function is used to return the coordinates of the polygons from
 a map in SpatialPolygons or SpatialPolygonsDataFrame formats.
}

\usage{
 abmap()
}


\details{
'abmap' returns a list with the coordinates of the municipalities
in the province of Albacete (Castilla-La mancha, Spain).
}

\examples{
require(opencpu)
opencpu$browse('/library/GeogebraR/www/albacete.html')
}

\keyword{interface}

