\name{albacete}
\alias{abmap}

\title{Example on maps}

\description{
 These functions are used in the example included in qcc.html on
Quality Control Charts.
}

\usage{
 abmap()
}

\arguments{
}

\details{
'abmap' returns a list with the coordinates of the municipalities
in the province of Albacete (Castilla-La mancha, Spain).
}

\examples{
require(opencpu)
opencpu$browse('/library/GeogebraR/www/albacete.html')
}

\keyword{interface}

