\name{fitcurves}
\alias{fitcurves}
\alias{getdata}

\title{Fit non-linear curves to data}

\description{
These functions are used in the example that fits a lowess, smooth spline
and 5th degree polynomial to a set of points.
}

\usage{
getdata (dataset)
fitcurves (xx)
}

\arguments{
\item{dataset}{Dataset to load. It can be one of "cars", "BOD" or "women".}
\item{xx}{A two-column matrix with the coordiantes of the data points.}
}

\keyword{interface}

