\name{granada}
\docType{data}
\alias{granada}
\title{Boundaries of the municipalities in the province of Granada (Spain)}
\description{
 This data set contains the boundaries of the municipalities in the province
 of Granada (Andalucía, Spain). 
}
\usage{granada}
\format{A SpatialPolygonsDataFrame with the boundaries and some additional}
\source{ Instituto Nacional de Estadística (España). \url{http://www.ine.es}}
\seealso{plotmap()}
\keyword{datasets}
