\name{jaen}
\docType{data}
\alias{jaen}
\title{Boundaries of the municipalities in the province of Jaén (Spain)}
\description{
 This data set contains the boundaries of the municipalities in the province
 of Jaén (Andalucía, Spain). 
}
\usage{jaen}
\format{A SpatialPolygonsDataFrame with the boundaries and some additional}
\source{ Instituto Nacional de Estadística (España). \url{http://www.ine.es}}
\seealso{plotmap()}
\keyword{datasets}
