\name{myronorm}
\alias{myrnorm}

\title{Simulate from rnorm().}

\description{
Wrapper function for rnorm() that allows for seting a seed.
}

\usage{
myrnorm(seed=NULL, ...)
}

\arguments{
\item{seed}{Seed to be passed to set.seed().}
\item{...}{Arguments to be passed to rnorm().}
}

\keyword{interface}

