\name{samplepp}
\alias{samplepp}

\title{Simulate point pattern of cases and controls}

\description{
 This function simulates pattern of cases and controls in the unit square and it computes the ratio of densities and contour lines.
}

\usage{
samplepp(ncases=10, ncontrols=30)
}

\arguments{
\item{ncases}{Number of cases.}
\item{ncontrols}{Number of controls.}
}

\keyword{interface}

