\name{simpts}
\alias{simpts}

\title{Simulate points.}

\description{
 This function simulates points using a standard normal for each coordinate.
}

\usage{
 simpts(n=1, seed=NULL)
}

\arguments{
\item{n}{Number of points to simulate.}
\item{seed}{Seed to be passed to set.seed().}
}

\keyword{interface}

