\name{albacete}
\docType{data}
\alias{albacete}
\title{Boundaries of the municipalities in the province of Albacete (Spain)}
\description{
 This data set contains the boundaries of the municipalities in the province
 of Albacete (Castilla-La Mancha, Spain). Albacete is also the city where 
 useR! 2013 took place.
}
\usage{albacete}
\format{A SpatialPolygonsDataFrame with the boundaries and some additional}
\source{ Instituto Nacional de Estadística (España). \url{http://www.ine.es}}
\seealso{abmap()}
\keyword{datasets}
