\name{CLT}
\alias{CLT}

\title{Central Limit Theorem example}

\description{
 This function simulates data and computes histogram, density and 
(asymptotic) density of the mean 
}

\usage{
 CLT(distr, params, nsims)
}

\arguments{
\item{distr}{Distribution to sample from.}
\item{params}{Vector of two elements with the parameters to use with 'distr'.}
\item{nsims}{Number of simulations.}
}

\keyword{interface}

